refactor: extract completion
