+++ ratisui-core/src/highlight_value.rs
@@ -436,7 +436,7 @@ fn get_node_path(node: Node) -> Vec<String> {
 
 #[cfg(test)]
 mod high_light_test {
-    use ratisui_core::highlight_value::HighlightKind;
+    use crate::highlight_value::HighlightKind;
     use anyhow::Result;
     use serde_json::json;
