+++ src/components/ft_search_panel.rs
@@ -1,0 +1,6 @@
+
+
+#[allow(unused)]
+pub struct FtSearchPanel {
+
+}
